domain code revise
