--- application/domain/IssuePriority.java
@@ -1,5 +1,0 @@
-package application.domain;
-
-public enum IssuePriority {
-    BLOCKER, CRITICAL, MAJOR, MINOR, TRIVIAL
-}
