--- application/domain/IssueStatus.java
@@ -1,5 +1,0 @@
-package application.domain;
-
-public enum IssueStatus {
-    NEW, ASSIGNED, RESOLVED, CLOSED, REOPENED
-}
