--- application/domain/UserRole.java
@@ -1,5 +1,0 @@
-package application.domain;
-
-public enum UserRole {
-    ADMIN, PROJECT_LEAD, DEVELOPER, TESTER
-}
