+++ application/src/main/java/com/se14/domain/Comment.java
@@ -1,15 +1,63 @@
-package com.se14.domain;
+package application.src.main.java.com.se14.domain;
 
-import lombok.Builder;
-import lombok.Data;
+import java.util.Date;
 
-import java.time.LocalDateTime;
-
-@Data // 빠른 개발을 위해 모든 Domain에 @Data 어노테이션을 추가함
-@Builder
 public class Comment {
-    private long id;
-    private String content;
-    private LocalDateTime timeStamp;
-    private Member author;
+    private String commentId;
+    private String text;
+    private Date timestamp;
+    private User author;
+
+    @Override
+    public String toString()
+    {
+        return this.commentId + ":" +this.text;
+    }
+    // Default constructor
+    public Comment() {
+    }
+
+    // Parameterized constructor
+    public Comment(String commentId, String text, Date timestamp, User author) {
+        this.commentId = commentId;
+        this.text = text;
+        this.timestamp = timestamp;
+        this.author = author;
+    }
+
+    // Getter and setter for commentId
+    public String getCommentId() {
+        return commentId;
+    }
+
+    public void setCommentId(String commentId) {
+        this.commentId = commentId;
+    }
+
+    // Getter and setter for text
+    public String getText() {
+        return text;
+    }
+
+    public void setText(String text) {
+        this.text = text;
+    }
+
+    // Getter and setter for timestamp
+    public Date getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(Date timestamp) {
+        this.timestamp = timestamp;
+    }
+
+    // Getter and setter for author
+    public User getAuthor() {
+        return author;
+    }
+
+    public void setAuthor(User author) {
+        this.author = author;
+    }
 }
