+++ application/src/main/java/com/se14/domain/IssuePriority.java
@@ -1,9 +1,5 @@
-package com.se14.domain;
+package application.src.main.java.com.se14.domain;
 
 public enum IssuePriority {
-    BLOCKER,
-    CRITICAL,
-    MAJOR,
-    MINOR,
-    TRIVIAL,
+    BLOCKER, CRITICAL, MAJOR, MINOR, TRIVIAL
 }
