+++ application/src/main/java/com/se14/domain/IssueStatus.java
@@ -1,9 +1,5 @@
-package com.se14.domain;
+package application.src.main.java.com.se14.domain;
 
 public enum IssueStatus {
-    NEW,
-    ASSIGNED,
-    RESOLVED,
-    CLOSED,
-    REOPENED,
+    NEW, ASSIGNED, RESOLVED, CLOSED, REOPENED
 }
