+++ application/src/main/java/com/se14/domain/UserRole.java
@@ -1,8 +1,5 @@
-package com.se14.domain;
+package application.src.main.java.com.se14.domain;
 
 public enum UserRole {
-    ADMIN,
-    PROJECT_LEAD,
-    DEVELOPER,
-    TESTER,
+    ADMIN, PROJECT_LEAD, DEVELOPER, TESTER
 }
